Fixed: millis overflow working correctly for receive string
There was an issue where receive_string when millis overran. I was
working fine for when there was a timeout, but if there was no timeout,
the function would still return false.

Problem fixed by correcting return from millis from uint16_t to
uint32_t.

+++ src/timer.h
@@ -14,7 +14,7 @@
  *          has started.
  * @note    Will roll-over after 49.7 days.
  */
-uint16_t millis(void);
+uint32_t millis(void);
 
 #endif
